requirements.txt
flask==2.2.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==1.5.3
numpy==1.24.3
tqdm==4.65.0
pdf2image==1.16.3
pillow==9.5.0
pytesseract==0.3.10
nltk==3.8.1
scikit-learn==1.2.2
matplotlib==3.7.1
argparse==1.4.0
concurrent-log-handler==0.9.24
requests==2.31.0 